Remove Serializable
This commit removes Serializable because it is untested and tricky to
maintain. Future pickling issue should be addressed by overwritting
__setstate__() and __getstate__().

This commit also fixes the bug that task goals fail to be retained
across pickling. This bug exposes an algorithms to more variants of
an environment than it should have. This bug does not breaks an
algorithm, but has an impact on the performance.