Fix object_grasped to 1 if object is grapsed
currently if the stick is grasped, sometimes it can
slip back and forth inside the gripper if it colides
with the handle when the agent tries to insert it into
the handle.

this leads to the object_grasp reward becoming less than 1
and issues with the final reward function, such as multiple
states taking on the same reward.

this changes that behavior by fixing object_grasp to 1 if the
gripper and tool are near eachother, the gripper is not completely
closed, meaning that it dropped the object, and that the object is
off the ground.

The diff is difficult to navigate because of autoformatter changes
but refer to line 203 of sawyer_stick_pull_v2.py for the actual diff.